pyproject.toml
[project]
name = "ai-content-writter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chainlit>=2.7.1.1",
    "ddgs>=9.5.4",
    "openai-agents>=0.2.9",
    "pymupdf>=1.26.3",
    "python-docx>=1.2.0",
    "reportlab>=4.4.3",
]
